Ed-Tech Agentic Core copy 2/requirements.txt
streamlit
anthropic
pandas
openpyxl
python-dotenv
tiktoken
watchdog
